[#31] Extract enums from ast
