+++ include/SCC/ast/ast.h
@@ -5,80 +5,11 @@
 #include <utility>
 #include <vector>
 
+#include "SCC/ast/data_types.h"
+#include "SCC/ast/statement_types.h"
 #include "SCC/config/scc_config.h"
 
 #include "logger/log.hpp"
-
-enum StatementType {
-  EMPTY_TYPE,
-  Program,
-  query,
-  ddlStatement,
-  dmlStatement,
-
-  // DDL Statements
-  alterTableStatement,
-  createDatabaseStatement,
-  createTableStatement,
-  dropDatabaseStatement,
-  dropTableStatement,
-
-  // DDL Basic Statements
-  tableDefinition,
-  columnDefinition,
-  tableConstraint,
-  alterActionADD,
-  alterActionDROP,
-  dropList,
-  dropConstraint,
-  dropColumn,
-
-  // DML Statements
-  deleteStatement,
-  insertStatement,
-  updateStatement,
-
-  // DML Basic Statements
-  condition,
-  ORcondition,
-  ANDcondition,
-  NOTcondition,
-  predicate,
-  expression,
-
-  // Basic Statements
-  primaryKey,
-  foreignKey,
-  reference,
-  name,
-  identifier,
-  delimiter_dot,
-  delimiter_comma,
-  delimiter_semicolon,
-
-  // Other key words
-  kw_constraint,
-
-  // SQL datatypes
-  SQL_int,
-  SQL_float,
-  SQL_char,
-  SQL_varchar,
-
-  StTypeCount
-};
-
-enum DataType {
-  ROOT,
-  SERVICE,
-  INT_NUMBER,
-  FLOAT_NUMBER,
-  BRACKET,
-  PUNCTUATION,
-  OPERATOR,
-  WORD,
-  STRING
-};
 
 class Node {
 public:
